initial work in refactoring by transfering the functions in modules.utils to clean up autopwn.py and for easier transition to src/ layout
